+++ consensus/common/src/main/java/org/hyperledger/besu/consensus/common/bft/IbftProcessor.java
@@ -12,9 +12,8 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.consensus.ibft;
+package org.hyperledger.besu.consensus.common.bft;
 
-import org.hyperledger.besu.consensus.common.bft.EventMultiplexer;
 import org.hyperledger.besu.consensus.common.bft.events.IbftEvent;
 
 import java.util.Optional;
