+++ consensus/common/src/main/java/org/hyperledger/besu/consensus/common/bft/MessageTracker.java
@@ -12,7 +12,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.consensus.ibft;
+package org.hyperledger.besu.consensus.common.bft;
 
 import static java.util.Collections.newSetFromMap;
