+++ consensus/ibft/src/integration-test/java/org/hyperledger/besu/consensus/ibft/support/StubbedSynchronizerUpdater.java
@@ -16,7 +16,7 @@ package org.hyperledger.besu.consensus.ibft.support;
 
 import static java.util.function.Function.identity;
 
-import org.hyperledger.besu.consensus.ibft.SynchronizerUpdater;
+import org.hyperledger.besu.consensus.common.bft.SynchronizerUpdater;
 import org.hyperledger.besu.ethereum.p2p.rlpx.connections.PeerConnection;
 
 import java.util.Collection;
